[apps/web] files count
