+++ apps/web/components/MessageForm/FilesSummary/index.module.scss
@@ -4,3 +4,8 @@
   color: $color-gray-700;
   font-size: 12px;
 }
+
+.file {
+  color: $color-indigo-700;
+  cursor: pointer;
+}
